(#160) Added Background to the editor
