+++ data/GameObjects/__init__.py
@@ -3,3 +3,4 @@ from .Door import Door
 from .Plate import Plate
 from .SpawnPoint import SpawnPoint
 from .Player import Player
+from .Background import Background
