+++ data/Level/LevelManager.py
@@ -3,19 +3,6 @@ from pygame.locals import *
 from data.GameObjects import *
 from data.utils.SaveManager import load_map
 
-class Background(pygame.sprite.Sprite):
-    def __init__(self, window_size):
-        super().__init__()
-        background_img = pygame.image.load('Resources/background.png').convert()
-        self.image = pygame.transform.scale(background_img, window_size)
-        self.rect = self.image.get_rect()
-
-    def on_collision(self, collider):
-        return
-
-    def has_collision(self, player_id):
-        return False
-
 class LevelManager:
     def load_map(self, file_path = 'map'):
         colliders = load_map(file_path)
